requirements.txt
# OpenClaw - Autonomous Agent Swarm
# Core
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
httpx==0.26.0

# LLM
google-generativeai==0.5.0

# Database (Shared Memory)
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Browser Automation
playwright==1.41.0

# Social Media APIs
tweepy==4.14.0         # Twitter/X
praw==7.7.1            # Reddit

# Utilities
python-dotenv==1.0.0
apscheduler==3.10.4    # Job scheduling
aiofiles==23.2.1
jinja2==3.1.3

# Google Cloud
google-cloud-compute==1.14.0
google-cloud-storage==2.14.0
